requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Login==0.6.3
python-dotenv==1.0.1
mysql-connector-python==9.1.0
Pillow==11.1.0
Flask-WTF==1.2.1
reportlab==4.0.8
Flask-Migrate==4.0.7
pandas==2.2.3
openpyxl==3.1.5
qrcode==8.0
